added inteerface for controller
Former-commit-id: 09171d4e56ad2bc79c9cde8d40a63352d5f0dd80

+++ src/gameengine/grid/interfaces/ActorGrid/ReadableGrid.java
@@ -2,13 +2,12 @@ package gameengine.grid.interfaces.ActorGrid;
 
 import java.util.Collection;
 
-import gameengine.actors.management.Actor;
 import gameengine.grid.interfaces.Identifiers.Grid2D;
 
 public interface ReadableGrid{
 	
 	Collection<Grid2D> getEnemiesInRadius(double x, double y, double radius);
-	Grid2D getLocationOf(double id);
+	Grid2D getLocationOf(int id);
 	Collection<Grid2D> getEnemyLocations();
 	Collection<Grid2D> getTowerLocations();
 	Collection<Grid2D> getBaseLocations();
